+++ fennec/apps/metamodel/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 
+from fennec.apps.metamodel.models import Change
 from fennec.apps.metamodel.utils import Index, ForeignKey, Schema, Layer, TableElement, RelationshipElement, Diagram
 from fennec.apps.metamodel.utils import Table, Namespace, Column
 
@@ -86,6 +87,7 @@ class ForeignKeyBasicSerializer(serializers.Serializer):
     onUpdate = serializers.IntegerField(source='on_update_referential_action')
     onDelete = serializers.IntegerField(source='on_delete_referential_action')
     tableRef = serializers.CharField(source='table_ref')
+
 
 class TableSerializer(serializers.Serializer):
     id = serializers.CharField()
@@ -229,3 +231,15 @@ class SandboxBasicInfoSerializer(serializers.Serializer):
     schemas = SchemaSerializer(many=True)
 
 
+class ChangeSerializer(serializers.ModelSerializer):
+    id = serializers.CharField(required=False)
+    content = serializers.CharField()
+    objectType = serializers.CharField(source="object_type")
+    objectCode = serializers.CharField(source="object_code")
+    changeType = serializers.IntegerField(source="change_type")
+    isUIChange = serializers.BooleanField(source="is_ui_change")
+    madeBy = serializers.IntegerField(source="made_by")
+
+    class Meta:
+        model = Change
+        fields = ('id', 'content', 'objectType', 'objectCode', 'changeType', 'isUIChange')
