+++ fennec/apps/repository/serializers.py
@@ -3,8 +3,6 @@ from django.contrib.auth.models import User, Group
 
 from fennec.apps.repository.models import Project, Branch,  BranchRevision, Sandbox
 
-from fennec.apps.metamodel.models import Change
-
 
 class UserSerializer(serializers.HyperlinkedModelSerializer):
     url = serializers.HyperlinkedIdentityField(view_name='user-detail', lookup_field='id')
@@ -55,17 +53,3 @@ class SandboxSerializer(serializers.ModelSerializer):
     class Meta:
         model = Sandbox
         fields = ('id',)
-
-
-class ChangeSerializer(serializers.ModelSerializer):
-    id = serializers.CharField(required=False)
-    content = serializers.CharField()
-    objectType = serializers.CharField(source="object_type")
-    objectCode = serializers.CharField(source="object_code")
-    changeType = serializers.IntegerField(source="change_type")
-    isUIChange = serializers.BooleanField(source="is_ui_change")
-    madeBy = serializers.IntegerField(source="made_by")
-
-    class Meta:
-        model = Change
-        fields = ('id', 'content', 'objectType', 'objectCode', 'changeType', 'isUIChange')
