+++ fennec/apps/repository/views.py
@@ -6,7 +6,8 @@ from django.contrib.auth.models import User, Group
 
 from fennec.apps.constants import MASTER_BRANCH_NAME, MASTER_BRANCH_TYPE, MASTER_BRANCH_DESCRIPTION
 
-from fennec.apps.metamodel.serializers import SchemaSerializer, DiagramSerializer, SandboxBasicInfoSerializer
+from fennec.apps.metamodel.serializers import SchemaSerializer, DiagramSerializer, SandboxBasicInfoSerializer, \
+    ChangeSerializer
 from fennec.apps.repository import utils
 from fennec.apps.repository.models import Project, Branch, BranchRevision, SandboxChange
 from fennec.apps.repository.serializers import BranchRevisionSerializer, ChangeSerializer
